[REF] web: basicRelationalModel: factorize legacyUtilities
The mapping operations between the new RelationalModel API and the
BasicModel is now done in its own file
